add security

+++ src/main/java/ru/sumarokov/task_tracker/service/TaskService.java
@@ -4,6 +4,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.sumarokov.task_tracker.entity.Task;
 import ru.sumarokov.task_tracker.exception.EntityNotFoundException;
+import ru.sumarokov.task_tracker.exception.NotAccessException;
+import ru.sumarokov.task_tracker.repository.TaskGroupRepository;
 import ru.sumarokov.task_tracker.repository.TaskRepository;
 
 import java.util.List;
@@ -12,25 +14,47 @@ import java.util.List;
 public class TaskService {
 
     private final TaskRepository taskRepository;
+    private final TaskGroupRepository taskGroupRepository;
+    private final AuthService authService;
 
     @Autowired
-    public TaskService(TaskRepository taskRepository) {
+    public TaskService(TaskRepository taskRepository, TaskGroupRepository taskGroupRepository, AuthService authService) {
         this.taskRepository = taskRepository;
+        this.authService = authService;
+        this.taskGroupRepository = taskGroupRepository;
     }
 
-    public List<Task> getTasks() {
-        return taskRepository.findAllByOrderByIdAsc();
+    public List<Task> getUserTasks() {
+        return taskRepository.findByTaskGroupUserOrderByIdAsc(authService.getUser());
     }
 
     public Task getTask(Long id) {
-        return taskRepository.findById(id).orElseThrow(EntityNotFoundException::new);
+        Task task = taskRepository.findById(id)
+                .orElseThrow(EntityNotFoundException::new);
+        if (task.getTaskGroup().getUser() != authService.getUser())
+            throw new NotAccessException();
+        return task;
     }
 
-    public Task saveTask(Task task) {
-        return taskRepository.save(task);
+    public void saveTask(Task task) {
+        if (task.getId() == -1L) {
+            taskRepository.save(task);
+        } else {
+            Task oldTask = taskRepository.findById(task.getId())
+                    .orElseThrow(EntityNotFoundException::new);
+            if (oldTask.getTaskGroup().getUser() != authService.getUser()) {
+                throw new NotAccessException();
+            } else {
+                taskRepository.save(task);
+            }
+        }
     }
 
     public void deleteTask(Long id) {
+        Task task = taskRepository.findById(id)
+                .orElseThrow(EntityNotFoundException::new);
+        if (task.getTaskGroup().getUser() != authService.getUser())
+            throw new NotAccessException();
         taskRepository.deleteById(id);
     }
 }
